v4.0.4
- **FileFormat Anycubic:**
  - (Fix) Preview marker size and `PropertyFields` for given file version
  - (Fix) Do not cache and reuse layer images if equal on file write, as that is causing "file damage" message (#782, #787)
- **Tools:**
  - **Mask:**
    - (Improvement) Apply the mask with Multiply instead of BitwiseAnd (#747)
  - **Layer and pixel arithmetic**:
    - (Improvement) Use the byte scale value (1/255) for multiply operations
  - **Infill:**
    - (Add) Floor/Ceil thickness (#785)
    - (Fix) Remove the debug blocking window when selecting the honeycomb pattern
- (Upgrade) .NET from 6.0.23 to 6.0.24

+++ UVtools.Core/Extensions/DrawingExtensions.cs
@@ -6,9 +6,9 @@ namespace UVtools.Core.Extensions;
 public static class DrawingExtensions
 {
     public static Color FactorColor(this Color color, byte pixelColor, byte min = 0, byte max = byte.MaxValue) =>
-        FactorColor(color, pixelColor / 255f, min, max);
+        FactorColor(color, pixelColor / 255.0, min, max);
 
-    public static Color FactorColor(this Color color, float factor, byte min = 0, byte max = byte.MaxValue)
+    public static Color FactorColor(this Color color, double factor, byte min = 0, byte max = byte.MaxValue)
     {
         byte r = (byte)(color.R == 0 ? 0 :
             Math.Min(Math.Max(min, color.R * factor), max));
